requirements.txt
# Основная библиотека для Telegram-бота (версия 3.x — актуальная на 2025 год)
aiogram==3.13.1

# Для работы с переменными окружения (.env файл)
python-dotenv==1.0.1

# Для красивой работы с датами (необязательно, но удобно)
python-dateutil==2.9.0.post0

# Для асинхронных таймеров и планировщиков (Pomodoro, напоминания)
aiofiles==24.1.0

# Опционально: если захотите красивые эмодзи в логах или текстах
emoji==2.14.0

# Опционально: для удобного логирования (чтобы видеть в консоли, что делает бот)
loguru==0.7.2

# Опционально: если в будущем захотите добавить AI (например, Grok, GPT и т.д.)
# openai==1.51.0
# httpx==0.27.2